+++ db/migrate/1418311331_create_hourglass_time_bookings.rb
@@ -1,6 +1,4 @@
-require 'hourglass/hourglass_migration'
-
-class CreateHourglassTimeBookings < HourglassMigration
+class CreateHourglassTimeBookings < Hourglass::Migration
   def change
     create_table :hourglass_time_bookings do |t|
       t.datetime :start, null: false