+++ db/migrate/20190307074023_change_hourglass_time_log_comments_limit.rb
@@ -1,6 +1,4 @@
-require 'hourglass/hourglass_migration'
-
-class ChangeHourglassTimeLogCommentsLimit < HourglassMigration
+class ChangeHourglassTimeLogCommentsLimit < Hourglass::Migration
   def up
     change_column :hourglass_time_logs, :comments, :string, limit: 1024
   end