+++ lib/hourglass/assets.rb
@@ -13,7 +13,7 @@ class Hourglass::Assets < Sprockets::Environment
         env.append_path path
       end
       if Rails.env.production?
-        env.js_compressor = :uglify
+        env.js_compressor = Uglifier.new(harmony: true)
         env.css_compressor = :scss
       end
     end
@@ -47,6 +47,13 @@ class Hourglass::Assets < Sprockets::Environment
       File.join 'plugin_assets', Hourglass::PLUGIN_NAME.to_s
     end
 
+    def redmine_base_path
+      (Redmine::Utils.relative_url_root || '/')
+    end
+    def plugin_base_path
+      File.join(redmine_base_path, assets_directory_path)
+    end
+
     def asset_directory_map
       {
           javascript: 'javascripts',
@@ -59,11 +66,25 @@ class Hourglass::Assets < Sprockets::Environment
     end
 
     def path(path, options = {})
-      if options[:type].present?
-        File.join '..', Rails.env.production? ? instance.find_asset(path).digest_path : File.join(asset_directory_map[options[:type]] || '', path)
+      resolved = instance.find_asset(path, options)
+      if resolved.nil?
+        if path.starts_with? plugin_base_path
+          path
+        else
+          redmine_path path, options
+        end
       else
-        path
+        return File.join(plugin_base_path, resolved.digest_path) if Rails.env.production?
+        plugin_path path, options
       end
+    end
+
+    def plugin_path(path, options = {})
+      File.join(plugin_base_path, asset_directory_map[options[:type]] || '', path)
+    end
+
+    def redmine_path(path, options = {})
+      File.join(redmine_base_path, asset_directory_map[options[:type]] || '', path)
     end
   end
 end