+++ lib/hourglass/hourglass_migration.rb
@@ -1,2 +1,0 @@
-class HourglassMigration < (Rails::VERSION::MAJOR <= 4 ? ActiveRecord::Migration : ActiveRecord::Migration[4.2])
-end